Added new term test file and added tests to term functions to that file.
